+++ src/components/Roadmap/RoadmapList/MyRoadmap/MyRoadmapHeader/style.ts
@@ -5,29 +5,19 @@ export const Header = styled.header`
   justify-content: space-between;
   align-items: center;
   flex-wrap: wrap-reverse;
-
-  & > h2 {
-    font-size: 28px;
-
-    @media ${({ theme }) => theme.mediaQuery.sm} {
-      font-size: 20px;
-    }
-  }
 `;
 
 export const ButtonContainer = styled.section`
   display: flex;
-  gap: 1rem;
+  gap: 16px;
 
-  @media ${({ theme }) => theme.mediaQuery.sm} {
-    justify-content: space-around;
+  @media ${({ theme }) => theme.mediaQuery.md} {
     width: 100%;
     margin-bottom: 15px;
 
     & > button {
       width: 100%;
       padding: 8px;
-      font-size: 14px;
     }
   }
 `;
