+++ src/pages/roadmap/index.tsx
@@ -12,13 +12,13 @@ interface RoadmapListProps {
 
 const RoadmapList = ({ isUserLogin }: RoadmapListProps) => {
   return (
-    <Root>
+    <>
       {isUserLogin ? <GNB /> : <GuestGNB />}
       <RoadmapListPage>
         {isUserLogin && <MyRoadmap />}
         <RecruitRoadmap />
       </RoadmapListPage>
-    </Root>
+    </>
   );
 };
 
@@ -42,19 +42,16 @@ export const getServerSideProps: GetServerSideProps<RoadmapListProps> = async (c
   };
 };
 
-const Root = styled.div``;
-
 const RoadmapListPage = styled.main`
   display: flex;
   flex-direction: column;
-  gap: 2rem;
+  gap: 15px;
   max-width: 1120px;
+  margin: 30px auto 0;
   padding: 0 40px;
-  margin: 40px auto 0;
 
-  @media ${({ theme }) => theme.mediaQuery.sm} {
-    margin: 20px auto;
-    padding: 0 10px;
-    gap: 15px;
+  @media ${({ theme }) => theme.mediaQuery.md} {
+    padding: 0 15px;
+    margin-top: 15px;
   }
 `;
