+++ src/styles/globals.css
@@ -65,6 +65,10 @@ h6 {
   font-weight: 600;
 }
 
+h1 {
+  font-size: 30px;
+}
+
 @media screen and (max-width: 768px) {
   h1 {
     font-size: 26px;
